backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-cors-headers==2.4.0
django-extensions==2.1.3
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
oauthlib==2.1.0
openapi-codec==1.3.2
pandas==0.23.4
pandas-drf-tools==0.1.1
Pillow==5.3.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
